pyproject.toml
[project]
name = "vpc-reporter"
version = "0.1.0"
description = "Comprehensive AWS VPC network documentation tool with beautiful CLI output"
readme = "README.md"
authors = [
    { name = "Asim Ali", email = "s.asim.ali.se@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "click>=8.1.0",
    "rich>=13.7.0",
    "boto3>=1.34.0",
    "aioboto3>=12.0.0",
    "aiofiles>=23.2.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "pyyaml>=6.0",
    "diskcache>=5.6.0",
    "loguru>=0.7.0",
    "typing-extensions>=4.8.0",
    "diagrams>=0.23.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "moto[all]>=4.2.0",
    "black>=23.12.0",
    "ruff>=0.1.8",
    "mypy>=1.8.0",
    "types-pyyaml>=6.0.0",
    "types-aiofiles>=23.2.0",
]

[project.scripts]
awsvpc = "vpc_reporter.cli.main:cli"

[build-system]
requires = ["uv_build>=0.8.19,<0.9.0"]
build-backend = "uv_build"

[tool.black]
line-length = 88
target-version = ["py311"]
include = '\.pyi?$'

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "C4", "UP"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["vpc_reporter"]
force-single-line = false
lines-after-imports = 2

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_generics = true
check_untyped_defs = true
no_implicit_reexport = true

[tool.pytest.ini_options]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "e2e: End-to-end tests",
    "aws: Tests requiring AWS credentials",
]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=vpc_reporter --cov-report=html --cov-report=term"
